Add sortGroup to list items
With this commit each item will have a sort group. Each group has a type and the value that has to be used for grouping. The actual grouping must be done by the client because the server doesn't have all the required information. For instance, grouping by date requires knowing the timezone of the user.