Add argument parsing and CLI mode
With this commit the app as a whole is a CLI. The "start" subcommand provides a way to start the various components of the app (for now there's just an "http-server" component, but there will be more in the future). The "cli" subcommand will allow for some management tasks to be performed (e.g. creating users, resetting password, etc.).

With this commit the "cli add user" subcommand has also been added, which allows to create a new user. Options can be passed either via CLI arguments or via an interactive mode if not provided.